Implement search-bar, music and comment section

+++ backend/src/main/java/com/oryanend/tom_perfeito_api/services/CommentService.java
@@ -27,18 +27,35 @@ public class CommentService {
   @Autowired private CommentRepository repository;
 
   @Transactional(readOnly = true)
-  public Page<CommentDTO> findAllPaged(Pageable pageable) {
-    Page<Comment> list = repository.findAll(pageable);
+  public Page<CommentDTO> findByMusic(UUID musicId, Pageable pageable) {
+
+    Music music =
+        musicRepository
+            .findById(musicId)
+            .orElseThrow(() -> new ResourceNotFoundException("Music not found"));
+
+    Page<Comment> list = repository.findByMusic(music, pageable);
+
     return list.map(CommentDTO::new);
   }
 
   @Transactional(readOnly = true)
-  public CommentDTO findById(Long id) {
-    Comment entity =
+  public CommentDTO findByIdAndMusic(Long commentId, UUID musicId) {
+    Music music =
+        musicRepository
+            .findById(musicId)
+            .orElseThrow(() -> new ResourceNotFoundException("Music not found"));
+
+    Comment comment =
         repository
-            .findById(id)
+            .findById(commentId)
             .orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
-    return new CommentDTO(entity);
+
+    if (!comment.getMusic().getId().equals(music.getId())) {
+      throw new ResourceNotFoundException("Comment does not belong to this music");
+    }
+
+    return new CommentDTO(comment);
   }
 
   @Transactional
@@ -83,9 +100,17 @@ public class CommentService {
   }
 
   @Transactional
-  public void delete(Long id) {
+  public void delete(Long id, UUID musicId) {
     try {
-      Comment comment = repository.getReferenceById(id);
+      Comment comment =
+          repository
+              .findById(id)
+              .orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
+
+      if (!comment.getMusic().getId().equals(musicId)) {
+        throw new ResourceNotFoundException("Comment does not belong to this music");
+      }
+
       authService.validateCreatedCommentBySelfOrAdmin(comment);
 
       repository.deleteById(id);
